Move CollaboratorKeyProvider to model package
